+++ src/main/java/com/Acknowledgement/Exceptions/AcknowledgementExistsException.java
@@ -1,4 +1,4 @@
-package com.Acknowledgement.exceptions;
+package com.Acknowledgement.Exceptions;
 
 public class AcknowledgementExistsException extends RuntimeException {
