+++ src/main/java/com/Acknowledgement/Exceptions/AcknowledgementNotFoundException.java
@@ -1,4 +1,4 @@
-package com.Acknowledgement.exceptions;
+package com.Acknowledgement.Exceptions;
 
 public class AcknowledgementNotFoundException extends RuntimeException {
