+++ src/main/java/com/User/Controller/UserController.java
@@ -1,0 +1,10 @@
+package com.User.Controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/users")
+public class UserController {
+
+}
