+++ src/main/java/com/User/Repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.User.repository;
+package com.User.Repository;
 
-import com.User.entity.User;
+import com.User.Entity.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
